refactor: rename files and add typescript
