+++ src/components/Header/components/Logo/Logo.tsx
@@ -1,8 +1,8 @@
-import React from 'react';
+import React, { FC } from 'react';
 import './Logo.css';
 import LogoImage from '../../../../assets/logo.png';
 
-const Logo = () => {
+const Logo: FC = () => {
 	return <img className='header__logo' src={LogoImage} alt='' />;
 };
